requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
Pillow==9.5.0
matplotlib==3.7.1
numpy==1.24.3
seaborn==0.12.2
plotly==5.14.1
